Only remove syllabification periods when syllabifying
When a sound change ends with ' x' (space followed by an 'x'), periods
are added between syllables. After the rule is applied, all periods are
removed. However, periods were being removed even if the sound change
did not add them. This commit adds a flas 'didSyllabify', ensuring that
periods are only removed when when they are added by a sound change to
mark syllables. Note that if there are periods that are not added by the
sound change, they are still removed; all this commit does is ensure
that if the sound change did **not** add periods, then preexisting
periods are not removed.